Commit Repositories

+++ IncidentSystem/src/main/java/com/example/IncidentSystem/Repository/BusinessRoleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface BusinessRoleRepository extends JpaRepository<BusinessRole, Long> {
+public interface BusinessRoleRepository extends JpaRepository<BusinessRole, Integer> {
 
 }
